Major updates and fixes
The mixin and classes now actually work and set background-color values
as well.

Added a partial to the base folder for improved importing from npm (no
need to go into the scss folder) and updated documentation to represent
the change.

Added inverse options in the mixin and classes.

Also, 51-102 are now available too!